Move InstrumentDescriptor into internal package
We use this class to track instrument sources/debugging so we don't want it exposed publicly.   This just moves it next to MetricDescriptor.
