+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/view/DefaultAggregation.java
@@ -6,9 +6,9 @@
 package io.opentelemetry.sdk.metrics.view;
 
 import io.opentelemetry.sdk.internal.ThrottlingLogger;
-import io.opentelemetry.sdk.metrics.common.InstrumentDescriptor;
 import io.opentelemetry.sdk.metrics.exemplar.ExemplarFilter;
 import io.opentelemetry.sdk.metrics.internal.aggregator.Aggregator;
+import io.opentelemetry.sdk.metrics.internal.descriptor.InstrumentDescriptor;
 import java.util.logging.Level;
 import java.util.logging.Logger;
