+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/view/NoAggregation.java
@@ -5,9 +5,9 @@
 
 package io.opentelemetry.sdk.metrics.view;
 
-import io.opentelemetry.sdk.metrics.common.InstrumentDescriptor;
 import io.opentelemetry.sdk.metrics.exemplar.ExemplarFilter;
 import io.opentelemetry.sdk.metrics.internal.aggregator.Aggregator;
+import io.opentelemetry.sdk.metrics.internal.descriptor.InstrumentDescriptor;
 
 /** Configuration representing no aggregation. */
 class NoAggregation extends Aggregation {
