+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/internal/descriptor/TestMetricDescriptor.java
@@ -7,7 +7,6 @@ package io.opentelemetry.sdk.metrics.internal.descriptor;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.opentelemetry.sdk.metrics.common.InstrumentDescriptor;
 import io.opentelemetry.sdk.metrics.common.InstrumentType;
 import io.opentelemetry.sdk.metrics.common.InstrumentValueType;
 import io.opentelemetry.sdk.metrics.view.View;
